% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{coef.GPPM}
\alias{coef.GPPM}
\title{Extract point estimates from a Gaussian process panel model}
\usage{
\method{coef}{GPPM}(model, ...)
}
\description{
Extract point estimates from a Gaussian process panel model
}
