% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{covf}
\alias{covf}
\title{Covariance Function}
\usage{
covf(gpModel, ...)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
The covariance function as a string.
}
\description{
Extracts the covariance function from a GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
myCov <- covf(simpleModel)
}
\seealso{
Other functions to extract from a GPPM: \code{\link{SE}},
  \code{\link{coef.GPPM}}, \code{\link{confint.GPPM}},
  \code{\link{datas}}, \code{\link{fitted.GPPM}},
  \code{\link{getIntern}}, \code{\link{logLik.GPPM}},
  \code{\link{maxntime}}, \code{\link{meanf}},
  \code{\link{nobs.GPPM}}, \code{\link{npar}},
  \code{\link{npred}}, \code{\link{nstime}},
  \code{\link{paramEsts}}, \code{\link{preds}},
  \code{\link{variable.names.GPPM}},
  \code{\link{vcov.GPPM}}
}
