% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoLGCM}
\alias{demoLGCM}
\title{Simulated Data From a Latent Growth Curve Model.}
\format{
A data frame with 1998 rows and 3 variables:
\describe{
\item{ID}{Subject ID}
\item{t}{Time index}
\item{y}{Generic measurement}
}
}
\usage{
demoLGCM
}
\description{
Simulated Data From a Latent Growth Curve Model.
}
\keyword{datasets}
