% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleModel}
\alias{exampleModel}
\title{Example Gaussian Process Panel model}
\format{An object of class GPPM. A fitted Gaussian process panel model. Used to speed up the example code for the numerous extractor functions. Result of\preformatted{data('demoLGCM')
exampleModel <- fit(gppm('muI+muS*t','varI+covIS*(t+t#)+varS*t*t#+(t==t#)*sigma',demoLGCM,'ID','y'))
}}
\usage{
exampleModel
}
\description{
Example Gaussian Process Panel model
}
\keyword{datasets}
