% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleModelNotFit}
\alias{exampleModelNotFit}
\title{Example Gaussian Process Panel model}
\format{An object of class GPPM. An unfitted Gaussian process panel model. Used for testing only. Result of\preformatted{data('demoLGCM')
exampleModelNotFit <- fit(gppm('muI+muS*t','varI+covIS*(t+t#)+varS*t*t#+(t==t#)*sigma',demoLGCM,'ID','y'))
}}
\usage{
exampleModelNotFit
}
\description{
Example Gaussian Process Panel model
}
\keyword{datasets}
