% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Generic Method For Fitting a model}
\usage{
fit(gpModel, ...)
}
\arguments{
\item{gpModel}{a model.}

\item{...}{additional arguments.}
}
\value{
A fitted model
}
\description{
Generic method for fitting a model.
}
\seealso{
\code{\link{fit.GPPM}}
}
