% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPPMControl.R
\name{gppmControl}
\alias{gppmControl}
\title{Define settings for a Gaussian process panel model}
\usage{
gppmControl(stanModel = TRUE)
}
\arguments{
\item{stanModel}{boolean. Should the corresponding stan model be created? Should only be set to FALSE for testing purposes.
Not creating the stan model makes model fitting impossible but saves a lot of time.}
}
\value{
Settings for a Gaussian process panel model in an object of class 'GPPMControl'
}
\description{
This function is used to specify the settings of a Gaussian process panel model generated by \code{\link{gppm}}.
}
\seealso{
\code{\link{gppm}}
}
