% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPPM.R
\name{gppModel}
\alias{gppModel}
\title{Define a Gaussian process panel model}
\usage{
gppModel(mFormula, cFormula, myData, ID, DV)
}
\arguments{
\item{mFormula}{character string. Contains the specification of the mean function. See details for more information.}

\item{cFormula}{character string. Contains the specification of the covariance function. See details for more information.}

\item{myData}{data frame. Contains the data to which the model is fitted. Must be in the long-format.}

\item{ID}{character string. Contains the column label in myData which describes the subject ID.}

\item{DV}{character string. Contains the column label in myData which contains the to be modeled variable.}
}
\value{
a (unfitted) Gaussian process panel model, that is an object of class 'GPPM'
}
\description{
This function is used to specify a Gaussian process panel model,
which can then be fit using \code{\link{fit}}.
}
\examples{
# Defintion of a latent growth curve model
data("demoLGCM")
lgcm <- gppModel('muI+muS*t','varI+covIS*(t+t#)+varS*t*t#+(t==t#)*sigma',
        demoLGCM,'ID','x')
}
\seealso{
\code{\link{fit.GPPM}} for how to fit a GPPM
}
