% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{logLik.GPPM}
\alias{logLik.GPPM}
\title{Extract Log-Likelihood from GPPMs}
\usage{
\method{logLik}{GPPM}(gpModel)
}
\arguments{
\item{gpModel}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}
}
\value{
Returns an object of class logLik. Attributs are: "df" (\strong{d}egrees of \strong{f}reedom; number of estimated parameters in the model) and nobs (number of persons in the model)
}
\description{
Compute the log-likelihood for a GPPM at the maximum likelihood parameter values.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
simpleModel <- fit(simpleModel)
ll <- logLik(simpleModel)
}
