% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{maxntime}
\alias{maxntime}
\title{Maximum Number of Observations per Person}
\usage{
maxntime(gpModel)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
Maximum number of observations as a numeric.
}
\description{
Extracts the maximum number of observations per person from a GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
numberTimePoints <- maxtime(simpleModel)
}
\seealso{
Other functions to extract from a GPPM: \code{\link{SE}},
  \code{\link{coef.GPPM}}, \code{\link{confint.GPPM}},
  \code{\link{covf}}, \code{\link{datas}},
  \code{\link{fitted.GPPM}}, \code{\link{getIntern}},
  \code{\link{logLik.GPPM}}, \code{\link{meanf}},
  \code{\link{nobs.GPPM}}, \code{\link{npar}},
  \code{\link{npred}}, \code{\link{nstime}},
  \code{\link{paramEsts}}, \code{\link{preds}},
  \code{\link{variable.names.GPPM}},
  \code{\link{vcov.GPPM}}
}
