% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{maxtime.GPPM}
\alias{maxtime.GPPM}
\title{Extract the Maximum Number of Observations per Person from a GPPM}
\usage{
\method{maxtime}{GPPM}(gpModel, ...)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
Maximum number of observations as a numeric.
}
\description{
Extracts the maximum number of observations per person from a GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
numberTimePoints <- maxtime(simpleModel)
}
