% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{maxtime}
\alias{maxtime}
\title{Extract the Maximum Number of Observations per Person from a Longitudinal Model}
\usage{
maxtime(object)
}
\arguments{
\item{object}{a longitudinal model}
}
\description{
Extracts the maximum number of Observations per Person from a longitudinal model.
}
\seealso{
\code{\link{maxtime.GPPM}}
}
