% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{meanf}
\alias{meanf}
\title{Mean Function}
\usage{
meanf(gpModel)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
The mean function as a character string.
}
\description{
Extracts the mean function from a GPPM.
}
\examples{
\dontrun{
data("demoLGCM")
lgcm <- gppm('muI+muS*t','varI+covIS*(t+t#)+varS*t*t#+(t==t#)*sigma',
        demoLGCM,'ID','y')
myMean <- meanf(lgcm)
}
}
\seealso{
Other functions to extract from a GPPM: \code{\link{SE}},
  \code{\link{coef.GPPM}}, \code{\link{confint.GPPM}},
  \code{\link{covf}}, \code{\link{datas}},
  \code{\link{fitted.GPPM}}, \code{\link{getIntern}},
  \code{\link{logLik.GPPM}}, \code{\link{maxnObs}},
  \code{\link{nObs}}, \code{\link{nPars}},
  \code{\link{nPers}}, \code{\link{nPreds}},
  \code{\link{parEsts}}, \code{\link{pars}},
  \code{\link{preds}}, \code{\link{vcov.GPPM}}
}
