% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{nobs.GPPM}
\alias{nobs.GPPM}
\title{Extract the Number of persons from a GPPM}
\usage{
\method{nobs}{GPPM}(gpModel, ...)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
Number of persons as a numeric.
}
\description{
Extracts the number of persons from a GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
simpleModel <- fit(simpleModel)
numberPersons <- nobs(simpleModel)
}
