% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{npar.GPPM}
\alias{npar.GPPM}
\title{Extract the Number of Parameters from a GPPM}
\usage{
\method{npar}{GPPM}(gpModel, ...)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
Number of parameters as a numeric.
}
\description{
Extracts the number of parameters from a GPPM
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
numberPersons <- nobs(simpleModel)
}
