% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{npred.GPPM}
\alias{npred.GPPM}
\title{Extract the Number of Predictors}
\usage{
\method{npred}{GPPM}(gpModel, ...)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
Number of predictors as numeric.
}
\description{
Extracts the number of predictors from a GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
nPreds <- npred(simpleModel)
}
