% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{npred}
\alias{npred}
\title{Extract the Number of Predictors}
\usage{
npred(object)
}
\arguments{
\item{object}{a  model}
}
\value{
Number of predictors as numeric.
}
\description{
Extracts the number of predictors from a model.
}
\seealso{
\code{\link{npred.GPPM}}
}
