% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{nstime.GPPM}
\alias{nstime.GPPM}
\title{Extract the Number of Observations from a GPPM}
\usage{
\method{nstime}{GPPM}(gpModel, ...)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
Number of observations for each person as a numeric vector. The corresponding IDs are in the IDs attribute.
}
\description{
Extracts the number of observations for each person from a GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
numberTimePoints <- nstime(simpleModel)

}
