% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{nstime}
\alias{nstime}
\title{Number of Observations}
\usage{
nstime(gpModel)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
Number of observations for each person as a numeric vector. The corresponding IDs are in the IDs attribute.
}
\description{
Extracts the number of observations for each person from a GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
numberTimePoints <- nstime(simpleModel)

}
\seealso{
Other functions to extract from a GPPM: \code{\link{SE}},
  \code{\link{coef.GPPM}}, \code{\link{confint.GPPM}},
  \code{\link{covf}}, \code{\link{datas}},
  \code{\link{fitted.GPPM}}, \code{\link{getIntern}},
  \code{\link{logLik.GPPM}}, \code{\link{maxntime}},
  \code{\link{meanf}}, \code{\link{nobs.GPPM}},
  \code{\link{npar}}, \code{\link{npred}},
  \code{\link{paramEsts}}, \code{\link{preds}},
  \code{\link{variable.names.GPPM}},
  \code{\link{vcov.GPPM}}
}
