% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{paramEsts.GPPM}
\alias{paramEsts.GPPM}
\title{Extract the Essential Parameter Estimation Results}
\usage{
\method{paramEsts}{GPPM}(gpModel, level = 0.95)
}
\arguments{
\item{gpModel}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}
}
\value{
The names of the parameters

A data.frame containing the estimated parameters, standard errorsm and the lower and upper bounds of the confidence intervals.
}
\description{
Extracts the essential parameter estimation results for a GPPM
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
parameters <- variable.names(simpleModel)

}
