% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{pars}
\alias{pars}
\title{Parameter Names}
\usage{
pars(gpModel)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\description{
Extracts the parameter names from a GPPM.
}
\seealso{
Other functions to extract from a GPPM: \code{\link{SE}},
  \code{\link{coef.GPPM}}, \code{\link{confint.GPPM}},
  \code{\link{covf}}, \code{\link{datas}},
  \code{\link{fitted.GPPM}}, \code{\link{getIntern}},
  \code{\link{logLik.GPPM}}, \code{\link{maxnObs}},
  \code{\link{meanf}}, \code{\link{nObs}},
  \code{\link{nPars}}, \code{\link{nPers}},
  \code{\link{nPreds}}, \code{\link{parEsts}},
  \code{\link{preds}}, \code{\link{vcov.GPPM}}
}
