% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLong.R
\name{plot.LongData}
\alias{plot.LongData}
\title{Plot a Long Data Frame}
\usage{
plot.LongData(myData, plotIds, by, ID, DV)
}
\arguments{
\item{myData}{longitudinal data frame; possibly of class 'LongData'.}

\item{plotIds}{vector of IDs for which the data should be printed. Can be left empty. Then 5 IDs are picked randomly.}

\item{by}{label of the variable on the x-axis.}

\item{ID}{label of the ID column.}

\item{DV}{label of the variable on the y-axis.}
}
\value{
a fitted Gaussian process panel model, which is an object of class 'GPPM'
}
\description{
This function is used to plot longitudinal data in data frame.
}
\examples{
data("demoLGCM")
plot(demoLGCM,plotIds=c(1,2,3),by='t',ID='ID',DV='y')
plot(demoLGCM,by='t',ID='ID',DV='y') #five random ids
lgcm <- gppm('muI+muS*t','varI+covIS*(t+t#)+varS*t*t#+(t==t#)*sigma',
        demoLGCM,'ID','y')
lgcmData <- datas(lgcm)
plot(lgcmData) #for data sets from a gppm by, ID, and DV are not necessary
}
