% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.GPPM}
\alias{predict.GPPM}
\title{GPPM predictions}
\usage{
\method{predict}{GPPM}(gpModel, newData, ...)
}
\arguments{
\item{gpModel}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}

\item{newData}{a data frame with the same column names as the data frame used for generating \code{gpModel} with \code{\link{gppm}}}
}
\value{
Predictions of the dependent variable for all rows in newData. Conditional predictions for all persons in newData that are also present
in the data used for fitting gpModel; unconditional predictions for others persons.
See examples for format.
}
\description{
Predict values based on a GPPM
}
