% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{preds.GPPM}
\alias{preds.GPPM}
\title{Extract the Predictors}
\usage{
\method{preds}{GPPM}(gpModel)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
The names of the predictors.
}
\description{
Extracts the predictors from a GPPM
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
myPreds <- preds(simpleModel)
}
