% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{preds}
\alias{preds}
\title{Extract the Predictors}
\usage{
preds(gpModel)
}
\arguments{
\item{object}{a model using predictors.}
}
\value{
The names of the predictors.
}
\description{
Extracts the predictors from a model.
}
\seealso{
\code{\link{preds.GPPM}}
}
