% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{SE.GPPM}
\alias{SE.GPPM}
\title{Calculate  Standard Errors for a GPPM}
\usage{
\method{SE}{GPPM}(gpModel, ...)
}
\arguments{
\item{gpModel}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}
}
\value{
Standard errors for all parameters as a named numeric vector.
}
\description{
Returns the standard errors of the parameters of a fitted GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
simpleModel <- fit(simpleModel)
paraEsts <- SE(simpleModel)
}
