% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{SE}
\alias{SE}
\title{Standard Errors}
\usage{
SE(object)
}
\arguments{
\item{object}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}
}
\value{
Standard errors for all parameters as a named numeric vector.
}
\description{
Returns the standard errors of the parameters of a fitted GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
simpleModel <- fit(simpleModel)
paraEsts <- SE(simpleModel)
}
\seealso{
Other functions to extract from a GPPM: \code{\link{coef.GPPM}},
  \code{\link{confint.GPPM}}, \code{\link{covf}},
  \code{\link{datas}}, \code{\link{fitted.GPPM}},
  \code{\link{getIntern}}, \code{\link{logLik.GPPM}},
  \code{\link{maxnObs}}, \code{\link{meanf}},
  \code{\link{nObs}}, \code{\link{nPars}},
  \code{\link{nPers}}, \code{\link{nPreds}},
  \code{\link{parEsts}}, \code{\link{pars}},
  \code{\link{preds}}, \code{\link{vcov.GPPM}}
}
