% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.GPPM}
\alias{summary.GPPM}
\title{Summarize a GPPM object}
\usage{
\method{summary}{ggplot}(x, ...)
}
\arguments{
\item{...}{other arguments ignored (for compatibility with generic)}

\item{object}{GPPM object to summarise}
}
\description{
Summarize a GPPM object
}
\examples{
data("demoLGCM")
                demoLGCM,'ID','x')
summary(lgcm)
}
