% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{truePars}
\alias{truePars}
\title{Parameters used for generating \code{\link{demoLGCM}}.}
\format{A parameter vector.}
\usage{
truePars
}
\description{
Parameters used for generating \code{\link{demoLGCM}}.
}
\keyword{datasets}
