% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{variable.names.GPPM}
\alias{variable.names.GPPM}
\title{Extract the Parameter names}
\usage{
\method{variable.names}{GPPM}(gpModel, ...)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
The names of the parameters.

The variable names of the predictors
}
\description{
Extracts the parameter names from a GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
parameters <- variable.names(simpleModel)
}
