% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{vcov.GPPM}
\alias{vcov.GPPM}
\title{Variance-Covariance Matrix}
\usage{
\method{vcov}{GPPM}(gpModel)
}
\arguments{
\item{gpModel}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}
}
\value{
A matrix of the estimated covariances between the parameter estimates. This has row and column names corresponding to the parameter names.
}
\description{
Returns the variance-covariance matrix of the parameters of a fitted GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
simpleModel <- fit(simpleModel)
paraEsts <- vcov(simpleModel)
}
\seealso{
Other functions to extract from a GPPM: \code{\link{SE}},
  \code{\link{coef.GPPM}}, \code{\link{confint.GPPM}},
  \code{\link{covf}}, \code{\link{datas}},
  \code{\link{fitted.GPPM}}, \code{\link{getIntern}},
  \code{\link{logLik.GPPM}}, \code{\link{maxntime}},
  \code{\link{meanf}}, \code{\link{nobs.GPPM}},
  \code{\link{npar}}, \code{\link{npred}},
  \code{\link{nstime}}, \code{\link{paramEsts}},
  \code{\link{preds}}, \code{\link{variable.names.GPPM}}
}
