% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{coef.GPPM}
\alias{coef.GPPM}
\title{Point Estimates}
\usage{
\method{coef}{GPPM}(object, ...)
}
\arguments{
\item{object}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}

\item{...}{additional arguments (currently not used).}
}
\value{
Point estimates for all parameters as a named numeric vector.
}
\description{
Extracts point estimates for all parameters from a fitted GPPM.
}
\examples{
data("exampleModel")
paraEsts <- coef(exampleModel)
}
\seealso{
Other functions to extract from a GPPM: \code{\link{SE}},
  \code{\link{confint.GPPM}}, \code{\link{covf}},
  \code{\link{datas}}, \code{\link{fitted.GPPM}},
  \code{\link{getIntern}}, \code{\link{logLik.GPPM}},
  \code{\link{maxnObs}}, \code{\link{meanf}},
  \code{\link{nObs}}, \code{\link{nPars}},
  \code{\link{nPers}}, \code{\link{nPreds}},
  \code{\link{parEsts}}, \code{\link{pars}},
  \code{\link{preds}}, \code{\link{vcov.GPPM}}
}
