% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{confint.GPPM}
\alias{confint.GPPM}
\title{Confidence Intervals for GPPM Parameters}
\usage{
\method{confint}{GPPM}(gpModel, parm, level = 0.95, ...)
}
\arguments{
\item{gpModel}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence limits for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in % (by default 2.5% and 97.5%).
}
\description{
Computes confidence intervals for one or more parameters in a fitted GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
simpleModel <- fit(simpleModel)
paraEsts <- confint(simpleModel)
}
