% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{crossvalidate}
\alias{crossvalidate}
\title{Cross-validation.}
\usage{
crossvalidate(gpModel, foldVector)
}
\arguments{
\item{gpModel}{object of class GPPM.}

\item{foldVector.}{A vector describing the foldstructure to use. For example, created by \code{\link{createLeavePersonsOutFolds}}}
}
\value{
Cross-validation estimates of the mean squared error (MSE) and the negative log-predictive probability (nLPP)
}
\description{
Performs cross-validation of a Gaussian process panel model.
}
\details{
The fold vector, must be a vector of length nrow(datas(gpModel)) of integers from 1 to k. If foldVector[i]=j, then data point i is assigned to fold j.
}
\examples{
data("demoLGCM")
lgcm <- gppm('muI+muS*t','varI+covIS*(t+t#)+varS*t*t#+(t==t#)*sigma',
        demoLGCM,'ID','y')
theFolds <- createLeavePersonsOutFolds(lgcm)
crossvalidate(lgcm,theFolds)
}
