% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.GPPM}
\alias{fit.GPPM}
\title{Fit a Gaussian process panel model}
\usage{
\method{fit}{GPPM}(gpModel, init = "random", useOptimizer = TRUE)
}
\arguments{
\item{gpModel}{object of class GPPM. The Gaussian proces spanel model to be fitted.}

\item{init}{string or named numeric vector. Used to specify the starting values for the parameters. Can either be the string 'random' (default) or a numeric vector startVal of starting values. Which value belongs to which parameter is determined by the names attribute of startVal. See also the example.}

\item{useOptimizer}{boolean. Should the optimizer be used or not? For false the (possibly random) starting values are returned as the maximum likelihood values.}
}
\value{
a fitted Gaussian process panel model, which is an object of class 'GPPM'
}
\description{
This function is used to fit a Gaussian process panel model,
which has been specified fit using \code{\link{gppm}}.
}
\examples{
#regular usage
data("demoLGCM")
linearChange <- gppm('muI+muS*t','(t==t#)*sigma',
        demoLGCM,'ID','x')
linearChange <- fit(linearChange)

#starting values as ML results
startVals <- c(10,1,10)
names(startVals) <- c('muI','muS','sigma')
linearChangeFake <- fit(linearChange,init=startVals,useOptimizer=FALSE)
stopifnot(identical(startVals,coef(linearChangeFake)))
}
