% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{fitted.GPPM}
\alias{fitted.GPPM}
\title{Extract fitted person-specific mean vectors and covariance matrices}
\usage{
\method{fitted}{GPPM}(gpModel, ..)
}
\arguments{
\item{gpModel}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}
}
\value{
Returns a list structure with mean and covariances matrices. See example.
}
\description{
A fitted GPPM implies a mean vector and a covariance matrix for each person. These are returned by this function.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
simpleModel <- fit(simpleModel)
meansCovs <- fitted(simpleModel)

person1Mean <- meansCovs$mean[[1]]
person1Cov <- meansCovs$cov[[1]]
person1ID <- meansCovs$ID[[1]]
}
