% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{npar}
\alias{npar}
\title{Extract Number of Parameters from a Model}
\usage{
npar(object)
}
\arguments{
\item{object}{a statistical model}
}
\value{
Number of parameters as a numeric.
}
\description{
Extracts the number of parameters from a model.
}
