% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{nstime}
\alias{nstime}
\title{Extract the Number of Observations from a Longitudinal Model}
\usage{
nstime(object)
}
\arguments{
\item{object}{a longitudinal model}
}
\value{
Number of observations for each person as a numeric vector.
}
\description{
Extracts the number of observations for each person from a longitudinal model.
}
\seealso{
\code{\link{nstime.GPPM}}
}
