% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{paramEsts}
\alias{paramEsts}
\title{Extract the Essential Parameter Estimation Results}
\usage{
paramEsts(object, ...)
}
\arguments{
\item{object}{statistical model.}
}
\value{
Essential parameter estimation results.
}
\description{
Extracts the essential parameter estimation results for a model
}
