% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.GPPM}
\alias{predict.GPPM}
\title{GPPM predictions}
\usage{
\method{predict}{GPPM}(gpModel, newData, ...)
}
\arguments{
\item{gpModel}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}

\item{newData}{a data frame with the same column names as the data frame used for generating \code{gpModel} with \code{\link{gppm}}}
}
\value{
predictions of the dependent variable for all rows of newData. Conditional predictions for all persons in newData that are also present
in the data used for fitting gpModel; unconditional predictions for others persons.
See examples for format.
}
\description{
Predict values based on a GPPM
}
\examples{
data("demoLGCM")
#remove all measurements from person 1 and the first form person 2
predIdx <- c(which(demoLGCM$ID==1),which(demoLGCM$ID==2)[1])
fitDemoLGCM <- demoLGCM(setdiff(1:nrow(demoLGCM),predIdx)

lgcm <- gppm('muI+muS*t','varI+covIS*(t+t#)+varS*t*t#+(t==t#)*sigma',
        demoLGCM,'ID','y')
predictions <- predict(lgcm,demoLGCM[predIdx])
predictions
}
