% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{SE}
\alias{SE}
\title{Calculate Standard Errors}
\usage{
SE(object)
}
\arguments{
\item{object}{a fitted model object.}
}
\value{
Standard errors for all parameters as a named numeric vector.
}
\description{
Returns the standard errors of the main parameters of a fitted model object.
}
