% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.GPPM}
\alias{simulate.GPPM}
\title{XXXX}
\usage{
\method{simulate}{GPPM}(gpModel, parameterValues, seed = NULL)
}
\description{
XXXX
}
\examples{
data("demoLGCM")
lgcm <- gppm('muI+muS*t','varI+covIS*(t+t#)+varS*t*t#+(t==t#)*sigma',
        demoLGCM,'ID','x')

parameterValues <- c(10,-1,0,10,0,0.1)
names(parameterValues) <-c('muI','muS','varI','varS','covIS','sigma')
simData <- simulate(lgcm,parameterValues)
}
