% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.GPPM}
\alias{summary.GPPM}
\title{Summarizing GPPM}
\usage{
summary.GPPM(gpModel)
}
\value{
An object of classs "summary.GPPM", which is a list with 4 entries:
' \itemize{
  \item \code{modelSpecification} an object of class 'ModelSpecification' describing the specified models as a list with 4 entries
  \itemize{
      \item \code{meanFormula} formula for the mean function; output of \code{\link{meanf}}
      \item \code{covFormula} formula for the covariance function; output of \code{\link{covf}}
      \item \code{nPars} number of parameters; output of \code{\link{npar}}
      \item \code{params} parameter names; output of \code{\link{variable.names.GPPM}}
      \item \code{nPreds} number of predictors; output of \code{\link{npred}}
      \item \code{preds} predictors names; output of \code{\link{preds}}
  }
  \item Second item
}
}
\description{
Summarizing GPPM
}
