% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{variable.names.GPPM}
\alias{variable.names.GPPM}
\title{Parameter Names}
\usage{
\method{variable.names}{GPPM}(gpModel)
}
\arguments{
\item{gpModel}{object of class GPPM.}
}
\value{
The names of the paramters
}
\description{
Extracts the parameter names from a GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
parameters <- variable.names(simpleModel)
}
\seealso{
Other functions to extract from a GPPM: \code{\link{SE}},
  \code{\link{coef.GPPM}}, \code{\link{confint.GPPM}},
  \code{\link{covf}}, \code{\link{datas}},
  \code{\link{fitted.GPPM}}, \code{\link{getIntern}},
  \code{\link{logLik.GPPM}}, \code{\link{maxntime}},
  \code{\link{meanf}}, \code{\link{nobs.GPPM}},
  \code{\link{npar}}, \code{\link{npred}},
  \code{\link{nstime}}, \code{\link{paramEsts}},
  \code{\link{preds}}, \code{\link{vcov.GPPM}}
}
