% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{vcov.GPPM}
\alias{vcov.GPPM}
\title{Calculate Variance-Covariance Matrix for a GPPM}
\usage{
\method{vcov}{GPPM}(gpModel)
}
\arguments{
\item{gpModel}{object of class GPPM. Must be fitted, that is, a result from \code{\link{fit.GPPM}}.}
}
\value{
A matrix of the estimated covariances between the parameter estimates. This has row and column names corresponding to the parameter names.
}
\description{
Returns the variance-covariance matrix of the parameters of a fitted GPPM.
}
\examples{
data("demoLGCM")
simpleModel <- gppm('grandMean','(t==t#)*sigma',
        demoLGCM,'ID','y')
simpleModel <- fit(simpleModel)
paraEsts <- vcov(simpleModel)
}
